// max-most-frontend/src/Interfaces/Orders.ts
export interface HistoryData {
  created_by: string;
  direction: string;
  discount_amount: number;
  name: string;
  prescriber: string;
  price: number;
  quantity: number;
  sku: string;
  subtotal: number;
  total: number;
  website_order_date: string;
  website_order_id: number;
  website_prescription_id: number;
}

export interface HistoryResponse {
  readonly results: Array<HistoryData>;
  readonly page?: number;
  readonly count?: number;
  readonly total?: number;
  readonly pages?: number;
}